Pipeline seems to work ok now

+++ transcription-models/tools/datasets.py
@@ -7,6 +7,7 @@ from .utils import *
 from torch.utils.data import Dataset
 from abc import abstractmethod
 from random import randint
+from copy import deepcopy
 from tqdm import tqdm
 
 import torch.nn.functional as F
@@ -19,7 +20,7 @@ import os
 
 
 class TranscriptionDataset(Dataset):
-    def __init__(self, base_dir, splits, hop_length, data_proc, frame_length):
+    def __init__(self, base_dir, splits, hop_length, data_proc, frame_length, reset_data):
 
         # Check if the dataset exists in memory
         if not os.path.isdir(base_dir):
@@ -43,14 +44,15 @@ class TranscriptionDataset(Dataset):
         else:
             self.seq_length = max(self.data_proc.get_sample_range(self.frame_length))
 
-        os.makedirs(self.get_gt_dir(), exist_ok=True)
+        self.reset_data = reset_data
 
-        # TODO - parameterize removal
-        """
-        if os.path.exists(self.get_feats_dir()):
+        if os.path.exists(self.get_gt_dir()) and self.reset_data:
+            shutil.rmtree(self.get_gt_dir())
+            os.makedirs(self.get_gt_dir())
+
+        if os.path.exists(self.get_feats_dir()) and self.reset_data:
             shutil.rmtree(self.get_feats_dir())
-        os.makedirs(self.get_feats_dir())
-        """
+            os.makedirs(self.get_feats_dir())
 
         # Load the paths of the audio tracks
         self.tracks = []
@@ -72,11 +74,11 @@ class TranscriptionDataset(Dataset):
     def __getitem__(self, index):
         track = self.tracks[index]
 
-        data = self.data[track]
+        data = deepcopy(self.data[track])
 
         feats_path = self.get_feats_dir(track)
 
-        if os.path.exists(feats_path) and track != '00_Jazz3-150-C_comp':
+        if os.path.exists(feats_path):
             # TODO - different key for each module?
             feats = np.load(feats_path)['feats']
         else:
@@ -89,6 +91,7 @@ class TranscriptionDataset(Dataset):
         if self.frame_length is not None:
             # TODO - how much will it hurt to pad with zeros instead of actual previous/later frames?
 
+            # TODO - note splitting here for sample start
             sample_start = randint(0, len(data['audio']) - self.seq_length)
 
             frame_start = sample_start // self.hop_length
@@ -101,6 +104,7 @@ class TranscriptionDataset(Dataset):
             data['audio'] = data['audio'][sample_start : sample_end]
             data['tabs'] = data['tabs'][:, :, frame_start : frame_end]
             data['feats'] = feats[:, :, frame_start : frame_end]
+            #data.pop('notes')
 
         return data
 
@@ -123,7 +127,7 @@ class TranscriptionDataset(Dataset):
         return path
 
     def get_feats_dir(self, track=None):
-        path = os.path.join(GEN_DATA_DIR, self.dataset_name(), self.data_proc.get_name())
+        path = os.path.join(GEN_DATA_DIR, self.dataset_name(), self.data_proc.features_name())
         if track is not None:
             path = os.path.join(path, track + '.npz')
         return path
@@ -170,13 +174,14 @@ class MAESTRO(TranscriptionDataset):
 
 
 class GuitarSet(TranscriptionDataset):
-    def __init__(self, base_dir=None, splits=None, hop_length=512, data_proc=None, frame_length=None):
+    def __init__(self, base_dir=None, splits=None, hop_length=512,
+                 data_proc=None, frame_length=None, reset_data=False):
 
         self.base_dir = base_dir
         if self.base_dir is None:
             self.base_dir = os.path.join(HOME, 'Desktop', 'Datasets', self.dataset_name())
 
-        super().__init__(self.base_dir, splits, hop_length, data_proc, frame_length)
+        super().__init__(self.base_dir, splits, hop_length, data_proc, frame_length, reset_data)
 
     def get_tracks(self, split):
         jams_dir = os.path.join(self.base_dir, 'annotation')
@@ -207,8 +212,13 @@ class GuitarSet(TranscriptionDataset):
             tabs = load_jams_guitar_tabs(jams_path, self.hop_length, num_frames)
             data['tabs'] = tabs
 
+            i_ref, p_ref = load_jams_guitar_notes(jams_path)
+            p_ref = np.array([p_ref]).T
+            notes = np.concatenate((i_ref, p_ref), axis=-1)
+            data['notes'] = notes
+
             gt_path = self.get_gt_dir(track)
-            np.savez(gt_path, audio=audio, tabs=tabs)
+            np.savez(gt_path, audio=audio, tabs=tabs, notes=notes)
 
         data['track'] = track
 
@@ -237,3 +247,5 @@ class GuitarSet(TranscriptionDataset):
         # TODO - mirdata might be overkill if I don't use its load function
         # TODO - "flac" option which download flac instead
         mirdata.guitarset.download(data_home=save_dir)
+
+    # TODO - MusicNet already implemented
